requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
boto3>=1.34.0
pypdf2>=3.0.0
pdf2image>=1.17.0
python-dotenv>=1.0.0
